pyproject.toml
[tool.poetry]
name = "insurance-calculator"
version = "0.1.0"
description = ""
authors = ["Vasiliy08 <vasiliy.levin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"
uvicorn = "^0.32.0"
orjson = "^3.10.11"
punq = "^0.7.0"
python-multipart = "^0.0.17"
httpx = "^0.27.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.4"
pre-commit = "^4.0.1"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
polyfactory = "^2.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length=120
target-version='py312'
exclude=['_local.py', 'test.py']


[tool.ruff.lint]
select = [
    "B",    # flake8-bugbear
    "C",    # flake8-comprehensions
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "PTH",  # flake8-use-pathlib
    "UP",   # pyupgrade
    "W",    # pycodestyle warnings
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
    "PTH123", # open()
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
